[Clean code] Factorize step validation button
