+++ src/components/pages/definition/go.tsx
@@ -1,17 +1,9 @@
-import React, {useCallback} from 'react'
+import React from 'react'
 import {useTranslation, Trans} from 'react-i18next'
-import {Link} from 'react-router-dom'
 
-import {updateSteps, useDispatch} from 'store/actions'
-import {useProjectId} from 'store/selections'
-import {getPath} from 'store/url'
-
-import Button from 'components/button'
 import Layout from 'components/layout'
+import StepValidationButton from 'components/step_validation_button'
 
-const linkStyle: React.CSSProperties = {
-  textDecoration: 'none',
-}
 const contentStyle: React.CSSProperties = {
   color: colors.DARK_FOREST_GREEN,
   fontSize: 25,
@@ -33,16 +25,11 @@ const layoutContentStyle: React.CSSProperties = {
   flexDirection: 'column',
   justifyContent: 'space-between',
 }
+
 // This is a top level page and should never be nested in another one.
 // TOP LEVEL PAGE
 const GoPage = (): React.ReactElement => {
   const {t} = useTranslation()
-
-  const dispatch = useDispatch()
-  const projectId = useProjectId()
-  const handleClick = useCallback((): void => {
-    dispatch(updateSteps(projectId, {definition: true}))
-  }, [dispatch, projectId])
 
   return <Layout header={t('Définition')} bigTitle={t('Félicitations\u00A0!')} style={layoutStyle}>
     <div style={layoutContentStyle}>
@@ -58,9 +45,9 @@ const GoPage = (): React.ReactElement => {
         </Trans>
       </div>
       <div style={buttonContainerStyle}>
-        <Link to={getPath(['STEPS'], t)} style={linkStyle}>
-          <Button type="secondLevel" onClick={handleClick}>{t('C\'est parti\u00A0!')}</Button>
-        </Link>
+        <StepValidationButton stepId="definition">
+          {t('C\'est parti\u00A0!')}
+        </StepValidationButton>
       </div>
     </div>
   </Layout>
